Adds GIT_URL variable and option to pass ceph src with SHA1_OR_REF
A GIT_URL can be specified so that a custom git repo for ceph can be
given (override the default, which points to github).

Additionally, if a sha1 is given, the build process now checks first if
a working dir exists, in which case it cleans and checks out the code
instead of cloning the repo.